+++ utils/bubbleSort.hpp
@@ -1,6 +1,6 @@
 #ifndef BUBBLE_SORT_HPP
 #define BUBBLE_SORT_HPP
 
-void bubbleSort(int numberList[], int arrSize);
+void bubbleSort(double numberList[], int arrSize);
 
 #endif
